requirements.txt
streamlit==1.24.0
google-generativeai
pdf2image
python-dotenv